Move pattern matching to a separate file
